requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Task Queue
celery==5.3.6
redis==5.0.1

# Browser Automation
playwright==1.41.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# File Handling
python-multipart==0.0.6
pandas==2.1.4
openpyxl==3.1.2

# Templates
jinja2==3.1.3
aiofiles==23.2.1

# Configuration
pyyaml==6.0.1

# Development/Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
